Split darkmode logic and added toggle icon to all pages
